# [Do not merge until Git dependencies removed] Uncompressed Encoding for Affine Points