This PR implements a feature (`uncompressed_serde`) that optimizes the speed of serializing public parameters by using an uncompressed encoding that removes the overhead of computing the second coordinate of points. Note: this PR depends on PRs in `pasta_curves` and `neptune`, here: https://github.com/zcash/pasta_curves/pull/73 and https://github.com/lurk-lab/neptune/pull/192. For now, we point at branches on `lurk-lab` maintained forks.

I compared the performance on the `minroot` public parameters. Based on some eyeballing benchmarks run on my local machine, deserialization time reduces from 2.7s to 100ms. Downstream in [`lurk-rs`](https://github.com/lurk-lab/lurk-rs), where the circuit is much bigger -- and `serde` has become a significant time factor in initializing proving times -- deserialization time on the [`sha256`](https://github.com/lurk-lab/lurk-rs/blob/master/examples/sha256.rs) example reduces from 120s to 7s. Note, the uncompressed encoding for pasta points are `[u8; 64]`, so this does have the downside of eating twice as much memory compared to the normal/compressed encoding.

The significant improvement in deserialization time makes caching public parameters viable. This would short circuit the need to regenerate costly public parameters each run, improving the develop/iterate experience. This infrastructure currently exists in `lurk-rs`, and this PR mainly wants to improve its efficiency. Although hopefully, this clarifies a growing need for systematic approach to deal with public parameters in Nova, and leaves future work open to discussion.

As for the feature itself, it is implemented via the [`UncompressedEncoding`](https://docs.rs/group/0.13.0/group/trait.UncompressedEncoding.html) trait in `group`. We only apply the uncompressed serde to `CommitmentKey`, since the structure stores 90% of the data.