@CPerezz 2023-07-20
This will probably be superset by #198 

Also, if we plan to move to `halo2curves` we already provide this functionallity in our `SerdeObject` trait. See: https://github.com/privacy-scaling-explorations/halo2curves/blob/main/src/serde.rs