@srinathsetty 2023-09-27
Hi @CPerezz 

> This will probably be superset by #198
> 
> Also, if we plan to move to `halo2curves` we already provide this functionallity in our `SerdeObject` trait. See: https://github.com/privacy-scaling-explorations/halo2curves/blob/main/src/serde.rs

What is the best way to use SerdeObject trait to get fast deserialization of field elements? Are there examples of this somewhere to compose this automatically with serde's methods?