@huitseeker 2023-09-27
@srinathsetty some answers to these questions might reside in https://github.com/privacy-scaling-explorations/halo2curves/pull/39